+++ src/main/java/com/github/jlangch/aviron/examples/RealtimeScannerExample.java
@@ -31,6 +31,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
 
+import com.github.jlangch.aviron.Clamd;
 import com.github.jlangch.aviron.Client;
 import com.github.jlangch.aviron.FileSeparator;
 import com.github.jlangch.aviron.dto.ScanResult;
@@ -44,7 +45,6 @@ import com.github.jlangch.aviron.filewatcher.FileWatcher_FsWatch;
 import com.github.jlangch.aviron.filewatcher.FileWatcher_JavaWatchService;
 import com.github.jlangch.aviron.filewatcher.IFileWatcher;
 import com.github.jlangch.aviron.limiter.ClamdCpuLimiter;
-import com.github.jlangch.aviron.limiter.ClamdPid;
 import com.github.jlangch.aviron.limiter.CpuProfile;
 import com.github.jlangch.aviron.limiter.DynamicCpuLimit;
 import com.github.jlangch.aviron.limiter.ScheduledClamdCpuLimiter;
@@ -87,11 +87,11 @@ public class RealtimeScannerExample {
                                                 "06:00-21:59 @  50%",
                                                 "22:00-23:59 @ 100%"));
 
-            // replace the demo clamd pid file with your real one or pass clamd PID
-            final ClamdPid clamdPID = new ClamdPid(demoFS.getClamdPidFile());
+            // replace the demo clamd PID file with your real one or pass a clamd PID
+           final Clamd clamd = new Clamd(demoFS.getClamdPidFile());
 
             // setup the clamd CPU limiter
-            final ClamdCpuLimiter limiter = new ClamdCpuLimiter(clamdPID, new DynamicCpuLimit(everyday));
+            final ClamdCpuLimiter limiter = new ClamdCpuLimiter(clamd, new DynamicCpuLimit(everyday));
             limiter.mocking(MOCKING); // turn mocking on/off
             limiter.setClamdCpuLimitChangeListener(this::onCpuLimitChangeEvent);